requirements.txt
flask
gunicorn
pandas==2.0.3
numpy==1.23.5
scikit-learn
yfinance
tensorflow-cpu==2.12.0
curl_cffi